send-email: prompt-dependent exit codes
It seems very likely that most scripted callers would want to know when
(some) mails were interactively requested to be not sent, so indicate
this situation with a non-zero exit code. We use 10/11, because these
seem sufficiently distinct from "regular" error codes one would expect.

This is technically a backwards-incompatible behavior change, and
therefore would be safer to make opt-in. However, it is much easier to
imagine that a scripting user simply didn't consider the possibility of
an interactive cancellation, than that they really meant to ignore it.
Also, the damage resulting from reporting the situation too eagerly is
expected to be trivial, while the damage resulting from callers failing
to opt-in (which is obviously the status quo, and is likely to persist
in most cases) is potentially at least somewhat serious. This means that
making it opt-in has an opportunity cost, and I think the trade-off
favors making the new behavior unconditional.

For interactive calls from the command line, interactive cancellation is
arguably not unexpected, but a human user will be able to interpret the
somewhat unusual exit code in light of their immediately preceding
interactions just fine.

Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>